pyproject.toml
[tool.poetry]
name = "coursework-5"
version = "0.1.0"
description = ""
authors = ["alexandra kashina <kashin159@gmail.com>"]
readme = "README.md"

package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
redis = "^5.2.1"
black = "^24.10.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
flake8 = "^7.1.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.30.0"
coverage = "^7.6.9"
isort = "^5.13.2"
drf-yasg = "^1.21.8"
stripe = "^11.3.0"
celery = "^5.4.0"
eventlet = "^0.38.2"
django-celery-beat = "^2.7.0"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
